Add "Generate fuseji search key variants", "Max search key length for fuseji generation", "Max total fuseji count" and "Max consecutive fuseji count" dict/freq options
Add "Generate mazegaki search key variants" dict/freq option

Greatly reduced JL's peak memory usage while creating database files

+++ JL.Core/WordClass/JmdictWordClassUtils.cs
@@ -6,8 +6,10 @@ using JL.Core.Dicts;
 using JL.Core.Dicts.Interfaces;
 using JL.Core.Dicts.JMdict;
 using JL.Core.Frontend;
+using JL.Core.Japanese;
 using JL.Core.Utilities;
-using JL.Core.Utilities.Japanese;
+using JL.Core.Utilities.Database;
+using Microsoft.Data.Sqlite;
 
 namespace JL.Core.WordClass;
 
@@ -65,11 +67,42 @@ internal static class JmdictWordClassUtils
     internal static async Task Serialize()
     {
         Dictionary<string, List<JmdictWordClass>> jmdictWordClassDictionary = new(StringComparer.Ordinal);
+        Dict dict = DictUtils.SingleDictTypeDicts[DictType.JMdict];
+        bool useDB = dict.Options.UseDB.Value;
+        if (useDB && File.Exists(dict.DBPath))
+        {
+            // ReSharper disable once UseAwaitUsing
+            using SqliteConnection? connection = DBUtils.CreateDBConnectionForReadOnlyConnectionString(dict.ReadOnlyConnectionString);
+            Debug.Assert(connection is not null);
+            PopulateFromDB(jmdictWordClassDictionary, connection);
+        }
+        else
+        {
+            PopulateFromJmdictContents(jmdictWordClassDictionary);
+        }
+
+        string tempPartOfSpeechFilePath = PathUtils.GetTempPath(s_partOfSpeechFilePath);
+        FileStream fileStream = new(tempPartOfSpeechFilePath, FileStreamOptionsPresets.s_asyncCreate64KBufferFso);
+        await using (fileStream.ConfigureAwait(false))
+        {
+            await JsonSerializer.SerializeAsync(fileStream, jmdictWordClassDictionary, JsonOptions.s_jsoIgnoringWhenWritingNull).ConfigureAwait(false);
+        }
+
+        PathUtils.ReplaceFileAtomicallyOnSameVolume(s_partOfSpeechFilePath, tempPartOfSpeechFilePath);
+    }
+
+    private static void PopulateFromJmdictContents(Dictionary<string, List<JmdictWordClass>> jmdictWordClassDictionary)
+    {
         FrozenSet<string> validWordClasses = DeconjugatorUtils.ValidWordClasses;
 
         Dict dict = DictUtils.SingleDictTypeDicts[DictType.JMdict];
         foreach ((string key, IList<IDictRecord> jmdictRecordList) in dict.Contents)
         {
+            if (key.ContainsAny(JapaneseUtils.Fuseji))
+            {
+                continue;
+            }
+
             int jmdictRecordListCount = jmdictRecordList.Count;
             for (int i = 0; i < jmdictRecordListCount; i++)
             {
@@ -160,22 +193,186 @@ internal static class JmdictWordClassUtils
                         results.Add(record);
                     }
                 }
-
                 else
                 {
                     jmdictWordClassDictionary[key] = [record];
                 }
             }
         }
+    }
 
-        string tempPartOfSpeechFilePath = PathUtils.GetTempPath(s_partOfSpeechFilePath);
-        FileStream fileStream = new(tempPartOfSpeechFilePath, FileStreamOptionsPresets.s_asyncCreate64KBufferFso);
-        await using (fileStream.ConfigureAwait(false))
+    private static void PopulateFromDB(Dictionary<string, List<JmdictWordClass>> jmdictWordClassDictionary, SqliteConnection connection)
+    {
+        FrozenSet<string> validWordClasses = DeconjugatorUtils.ValidWordClasses;
+        connection.Open();
+
+        Dictionary<long, WordClassCandidate> rowIdToWordClassCandidate = [];
+        using SqliteCommand recordCommand = connection.CreateCommand();
+        recordCommand.CommandText =
+            $"""
+            SELECT {JmdictDBManager.RowId}, {JmdictDBManager.PrimarySpelling}, {JmdictDBManager.Readings}, {JmdictDBManager.PartOfSpeechSharedByAllSenses}, {JmdictDBManager.PartOfSpeech}
+            FROM {JmdictDBManager.Record}
+            """;
+
+        const int rowIdColumnIndex = 0;
+        const int primarySpellingColumnIndex = 1;
+        const int readingsColumnIndex = 2;
+        const int wordClassesSharedByAllSensesColumnIndex = 3;
+        const int wordClassesColumnIndex = 4;
+
+        using SqliteDataReader dataReader = recordCommand.ExecuteReader();
+        while (dataReader.Read())
         {
-            await JsonSerializer.SerializeAsync(fileStream, jmdictWordClassDictionary, JsonOptions.s_jsoIgnoringWhenWritingNull).ConfigureAwait(false);
+            string[]? wordClassesSharedByAllSenses = dataReader.GetNullableValueFromBlobStream<string[]>(wordClassesSharedByAllSensesColumnIndex);
+            string[]?[]? wordClasses = dataReader.GetNullableValueFromBlobStream<string[]?[]>(wordClassesColumnIndex);
+            if (wordClasses is null && wordClassesSharedByAllSenses is null)
+            {
+                continue;
+            }
+
+            List<string> wordClassList = [];
+            if (wordClasses is not null)
+            {
+                foreach (string[]? wordClassArray in wordClasses)
+                {
+                    if (wordClassArray is not null)
+                    {
+                        foreach (string wordClass in wordClassArray)
+                        {
+                            if (validWordClasses.TryGetValue(wordClass, out string? internedWordClass) && !wordClassList.Contains(internedWordClass))
+                            {
+                                wordClassList.Add(internedWordClass);
+                            }
+                        }
+                    }
+                }
+            }
+
+            if (wordClassesSharedByAllSenses is not null)
+            {
+                foreach (string wordClass in wordClassesSharedByAllSenses)
+                {
+                    if (validWordClasses.TryGetValue(wordClass, out string? internedWordClass) && !wordClassList.Contains(internedWordClass))
+                    {
+                        wordClassList.Add(internedWordClass);
+                    }
+                }
+            }
+
+            if (wordClassList.Count is 0)
+            {
+                continue;
+            }
+
+            long rowId = dataReader.GetInt64(rowIdColumnIndex);
+            string primarySpelling = dataReader.GetString(primarySpellingColumnIndex);
+            string[]? readings = dataReader.GetNullableValueFromBlobStream<string[]>(readingsColumnIndex);
+
+            string normalizedPrimarySpelling = JapaneseUtils.NormalizeText(primarySpelling);
+            string[]? normalizedReadings;
+            if (readings is not null)
+            {
+                normalizedReadings = new string[readings.Length];
+                for (int i = 0; i < readings.Length; i++)
+                {
+                    normalizedReadings[i] = JapaneseUtils.NormalizeText(readings[i]);
+                }
+            }
+            else
+            {
+                normalizedReadings = null;
+            }
+
+            rowIdToWordClassCandidate[rowId] = new WordClassCandidate(
+                primarySpelling,
+                normalizedPrimarySpelling,
+                readings,
+                normalizedReadings,
+                wordClassList.ToArray());
         }
 
-        PathUtils.ReplaceFileAtomicallyOnSameVolume(s_partOfSpeechFilePath, tempPartOfSpeechFilePath);
+        if (rowIdToWordClassCandidate.Count is 0)
+        {
+            LoggerManager.Logger.Error("PopulateFromDB failed unexpectedly, rowIdToRecord.Count is 0!");
+            return;
+        }
+
+        _ = jmdictWordClassDictionary.EnsureCapacity(rowIdToWordClassCandidate.Count);
+
+        using SqliteCommand searchKeyCommand = connection.CreateCommand();
+        searchKeyCommand.CommandText = $"SELECT {JmdictDBManager.RecordId}, {JmdictDBManager.SearchKey} FROM {JmdictDBManager.RecordSearchKey}";
+        const int recordIdColumnIndex = 0;
+        const int searchKeyColumnIndex = 1;
+
+        using SqliteDataReader searchKeyReader = searchKeyCommand.ExecuteReader();
+        while (searchKeyReader.Read())
+        {
+            long recordId = searchKeyReader.GetInt64(recordIdColumnIndex);
+            if (!rowIdToWordClassCandidate.TryGetValue(recordId, out WordClassCandidate? data))
+            {
+                continue;
+            }
+
+            string key = searchKeyReader.GetString(searchKeyColumnIndex);
+            if (key.ContainsAny(JapaneseUtils.Fuseji))
+            {
+                continue;
+            }
+
+            if (data.NormalizedReadings is not null)
+            {
+                bool keyFromReading = false;
+                foreach (string normalizedReading in data.NormalizedReadings)
+                {
+                    if (normalizedReading == key)
+                    {
+                        keyFromReading = true;
+                        break;
+                    }
+                }
+
+                if (keyFromReading)
+                {
+                    if (data.NormalizedPrimarySpelling != key)
+                    {
+                        continue;
+                    }
+
+                    if (jmdictWordClassDictionary.TryGetValue(key, out List<JmdictWordClass>? prevResults))
+                    {
+                        bool alreadyAdded = false;
+                        foreach (JmdictWordClass wordClass in prevResults.AsReadOnlySpan())
+                        {
+                            if (wordClass.Spelling == data.PrimarySpelling
+                                && wordClass.Readings.SequenceEqual(data.Readings)
+                                && wordClass.WordClasses.SequenceEqual(data.WordClasses))
+                            {
+                                alreadyAdded = true;
+                                break;
+                            }
+                        }
+
+                        if (alreadyAdded)
+                        {
+                            continue;
+                        }
+                    }
+                }
+            }
+
+            JmdictWordClass record = new(data.PrimarySpelling, data.WordClasses, data.Readings);
+            if (jmdictWordClassDictionary.TryGetValue(key, out List<JmdictWordClass>? results))
+            {
+                if (!results.AsReadOnlySpan().Contains(record))
+                {
+                    results.Add(record);
+                }
+            }
+            else
+            {
+                jmdictWordClassDictionary[key] = [record];
+            }
+        }
     }
 
     internal static async Task Initialize()
@@ -186,14 +383,14 @@ internal static class JmdictWordClassUtils
         if (!File.Exists(s_partOfSpeechFilePath)
             || (File.Exists(fullJmdictPath) && File.GetLastWriteTime(fullJmdictPath) > File.GetLastWriteTime(s_partOfSpeechFilePath)))
         {
-            bool useDB = jmdictDict.Options.UseDB.Value;
-            if (jmdictDict.Active && !useDB)
+            if (jmdictDict.Active)
             {
                 await Serialize().ConfigureAwait(false);
             }
             else
             {
                 bool deleteJmdictFile = false;
+                bool deleteDB = !File.Exists(jmdictDict.DBPath);
                 if (!File.Exists(fullJmdictPath))
                 {
                     deleteJmdictFile = true;
@@ -201,20 +398,26 @@ internal static class JmdictWordClassUtils
                     Uri? uri = jmdictDict.Url;
                     Debug.Assert(uri is not null);
 
-                    bool downloaded = await ResourceUpdater.DownloadBuiltInDict(fullJmdictPath,
-                        uri,
-                        jmdictDict.Type.ToString(), false, true).ConfigureAwait(false);
-
+                    bool downloaded = await ResourceUpdater.DownloadBuiltInDict(fullJmdictPath, uri, jmdictDict.Type.ToString(), false, true).ConfigureAwait(false);
                     if (!downloaded)
                     {
                         return;
                     }
                 }
 
-                jmdictDict.Contents = new Dictionary<string, IList<IDictRecord>>(jmdictDict.Size > 0 ? jmdictDict.Size : 450000, StringComparer.Ordinal);
+                bool useDB = jmdictDict.Options.UseDB.Value;
                 try
                 {
-                    await JmdictLoader.Load(jmdictDict).ConfigureAwait(false);
+                    if (useDB)
+                    {
+                        await JmdictDBManager.ImportFromDisk(jmdictDict).ConfigureAwait(false);
+                    }
+                    else
+                    {
+                        jmdictDict.Contents = new Dictionary<string, IList<IDictRecord>>(jmdictDict.Size > 0 ? jmdictDict.Size : JmdictLoader.Size, StringComparer.Ordinal);
+                        await JmdictLoader.Load(jmdictDict).ConfigureAwait(false);
+                    }
+
                     await Serialize().ConfigureAwait(false);
                 }
                 catch (Exception ex)
@@ -227,7 +430,17 @@ internal static class JmdictWordClassUtils
                     jmdictDict.Contents = FrozenDictionary<string, IList<IDictRecord>>.Empty;
                     if (deleteJmdictFile)
                     {
-                        File.Delete(fullJmdictPath);
+                        if (File.Exists(fullJmdictPath))
+                        {
+                            File.Delete(fullJmdictPath);
+                        }
+                    }
+                    if (deleteDB)
+                    {
+                        if (File.Exists(jmdictDict.DBPath))
+                        {
+                            File.Delete(jmdictDict.DBPath);
+                        }
                     }
                 }
             }
